+++ src/router/router.ts
@@ -9,7 +9,6 @@ import PasswordChanging from "../pages/PasswordChanging.vue";
 import LogInPage from "../pages/LogInPage.vue";
 import RegistrationPage from "../pages/RegistrationPage.vue";
 import SimpleSoundTest from "../pages/tests/sound/simple/SimpleSoundTest.vue";
-import SimpleSoundTest2 from "../pages/tests/sound/simple/SimpleSoundTest2.vue";
 
 const routes = [
   {
@@ -63,11 +62,7 @@ const routes = [
     path: '/test/sound/simple',
     component: SimpleSoundTest,
     meta: {ruName: "SoundSimpleTest"}
-  },  {
-    path: '/test/sound/simple2',
-    component: SimpleSoundTest2,
-    meta: {ruName: "SoundSimpleTest2"}
-  },
+  }
 
 ];
